requirements.txt
tensorflow
mediapipe
opencv-python
pillow
numpy
